Disabled sample (so we can publish a nuget package)
