--- Sample/Dog.cs
@@ -1,6 +1,0 @@
-namespace Sample;
-
-public class Dog
-{
-    public void MakeSound() => Console.WriteLine("Woof");
-}
