--- Sample/Duck.cs
@@ -1,6 +1,0 @@
-namespace Sample;
-
-public class Duck
-{
-    public void MakeSound() => Console.WriteLine("Quack");
-}
