--- Sample/IAnimal.cs
@@ -1,6 +1,0 @@
-namespace Sample;
-
-public interface IAnimal
-{
-    public void MakeSound();
-}
