+++ Sample/Sample.csproj
@@ -7,8 +7,4 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
-    <ItemGroup>
-      <PackageReference Include="MentalDesk.DuckType" Version="2.0.2" />
-    </ItemGroup>
-
 </Project>
